getFilterExpression: src/compile/data/filter => expression in src/filter
