+++ src/compile/data/filter.ts
@@ -1,8 +1,5 @@
-import {DateTime, dateTimeExpr, isDateTime} from '../../datetime';
-import {field} from '../../fielddef';
-import {isEqualFilter, isInFilter, isRangeFilter, Filter} from '../../filter';
-import {TimeUnit, fieldExpr as timeUnitFieldExpr, isSingleTimeUnit} from '../../timeunit';
-import {isArray, isString} from '../../util';
+import {expression} from '../../filter';
+import {isArray} from '../../util';
 
 import {FacetModel} from '../facet';
 import {LayerModel} from '../layer';
@@ -19,65 +16,17 @@ export namespace filter {
    * - a timestamp value of casted single time unit value
    * - stringified value
    */
-  function valueExpr(v: any, timeUnit: TimeUnit) {
-    if (isDateTime(v)) {
-      const expr = dateTimeExpr(v, true);
-      return 'time(' + expr + ')';
-    }
-    if (isSingleTimeUnit(timeUnit)) {
-      const datetime: DateTime = {};
-      datetime[timeUnit] = v;
-      const expr = dateTimeExpr(datetime, true);
-      return 'time(' + expr + ')';
-    }
-    return JSON.stringify(v);
-  }
-
-  export function getFilterExpression(filter: Filter | string) {
-    if (isString(filter)) {
-      return filter as string;
-    } else { // Filter Object
-      const fieldExpr = filter.timeUnit ?
-        // For timeUnit, cast into integer with time() so we can use ===, inrange, indexOf to compare values directly.
-          // TODO: We calculate timeUnit on the fly here. Consider if we would like to consolidate this with timeUnit pipeline
-          // TODO: support utc
-        ('time(' + timeUnitFieldExpr(filter.timeUnit, filter.field) + ')') :
-        field(filter, {datum: true});
-
-      if (isEqualFilter(filter)) {
-        return fieldExpr + '===' + valueExpr(filter.equal, filter.timeUnit);
-      } else if (isInFilter(filter)) {
-        return 'indexof([' +
-          filter.in.map((v) => valueExpr(v, filter.timeUnit)).join(',') +
-          '], ' + fieldExpr + ') !== -1';
-      } else if (isRangeFilter(filter)) {
-        const lower = filter.range[0];
-        const upper = filter.range[1];
-
-        if (lower !== null &&  upper !== null) {
-          return 'inrange(' + fieldExpr + ', ' +
-            valueExpr(lower, filter.timeUnit) + ', ' +
-            valueExpr(upper, filter.timeUnit) + ')';
-        } else if (lower !== null) {
-          return fieldExpr + ' >= ' + lower;
-        } else if (upper !== null) {
-          return fieldExpr + ' <= ' + upper;
-        }
-      }
-    }
-    return undefined;
-  }
 
   export function parse(model: Model): string {
     const filter = model.transform().filter;
     if (isArray(filter)) {
       return '(' +
-        filter.map((f) => getFilterExpression(f))
+        filter.map((f) => expression(f))
           .filter((f) => f !==undefined)
           .join(') && (') +
         ')';
     } else if (filter) {
-      return getFilterExpression(filter);
+      return expression(filter);
     }
     return undefined;
   }
